Zero-extend code units when casing them from 'signed char' to 'uint32_t'.
The problem starts with lexer: it should operate on unsigned chars
(as re2c wants). Then these casts won't be needed at all. Mixing
signed and unsigned chars is bad and lexer should be rewritten to
use unsigned chars.

+++ re2c/src/parse/unescape.cc
@@ -72,7 +72,7 @@ uint32_t unesc_oct (const char * s, const char * s_end)
 	for (++s; s != s_end; ++s)
 	{
 		n <<= 3;
-		n += static_cast<uint32_t> (*s - '0');
+		n += static_cast<uint8_t> (*s - '0');
 	}
 	return n;
 }
